requirements.txt
asn1crypto==0.23.0
astroid==1.5.3
backports.functools-lru-cache==1.4
bcrypt==3.1.3
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
cryptography==2.0.3
enum34==1.1.6
idna==2.6
ipaddress==1.0.18
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
paramiko==2.3.1
pyasn1==0.3.7
pycparser==2.18
pylint==1.7.4
pymssql==2.1.3
PyNaCl==1.1.2
pyOpenSSL==17.3.0
pypandoc==1.4
PyYAML==3.12
requests==2.18.4
robotframework==3.0.2
selenium==3.6.0
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.14
urllib3==1.22
wrapt==1.10.11
super_devops==0.0.8
